# trackingservice/target/classes/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: contact@egovernments.org
  description: "Tracking service provides client application the capability to create\
    \ routes, trips and track the trips. This service is also responsible for generating\
    \ alerts and notifications in case of special event."
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://sanitation.digit.org/
  title: Tracking service
  version: 1.0.11
...
servers:
- url: https://core.digit.org/api/v3
...
paths:
...
  /route/_update:
    put:
      description: Update an existing route by Id
      operationId: updateRoute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Route'
          application/xml:
            schema:
              $ref: '#/components/schemas/Route'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Route'
        description: Update an existent route in the system
        required: true
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid Route ID is supplied
        "404":
          description: Route not found
        "405":
          description: Validation exception
      summary: Update an existing route by id
      tags:
      - Route
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Route
  /route/_create:
    post:
      description: Create a new Route by Id
      operationId: createRoute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Route'
          application/xml:
            schema:
              $ref: '#/components/schemas/Route'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Route'
        description: Create a new Route in the system
        required: true
      responses:
        "200":
          description: Successful operation
        "405":
          description: Validation exception
      summary: Create a new route by providing a list of POIs
      tags:
      - Route
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Route
  /route/_search:
    get:
      description: Search Route based on multiple filters
      operationId: findRoute
      parameters:
      - description: Status values that need to be considered for filter
        explode: true
        in: query
        name: status
        required: false
        schema:
          default: active
          enum:
          - active
          - inactive
          type: string
        style: form
      - description: route name that needs to be considered for filter
        explode: true
        in: query
        name: routeName
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Route'
                type: array
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/Route'
                type: array
          description: successful operation
        "400":
          description: Invalid search value
      summary: Find Route based on supported parameters
      tags:
      - Route
      x-accepts: application/json
      x-tags:
      - tag: Route
  /route/{routeId}:
    get:
      description: Returns a single Route
      operationId: getRouteById
      parameters:
      - description: ID of Route to return
        explode: false
        in: path
        name: routeId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
            application/xml:
              schema:
                $ref: '#/components/schemas/Route'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Route not found
      summary: Find Route by its id
      tags:
      - Route
      x-accepts: application/json
      x-tags:
      - tag: Route
...
components:
  schemas:
...
    Audit:
      example:
        createdDate: 2000-01-23T04:56:07.000+00:00
        updatedBy: Id of the user who updated the entity
        createdBy: Id of the user who created the entity
        updatedDate: 2000-01-23T04:56:07.000+00:00
      properties:
        createdBy:
          example: Id of the user who created the entity
          type: string
        createdDate:
          format: date-time
          type: string
        updatedBy:
          example: Id of the user who updated the entity
          type: string
        updatedDate:
          format: date-time
          type: string
      type: object
      xml:
        name: audit
    Route:
      example:
        startPoi: startPoi
        audit:
          createdDate: 2000-01-23T04:56:07.000+00:00
          updatedBy: Id of the user who updated the entity
          createdBy: Id of the user who created the entity
          updatedDate: 2000-01-23T04:56:07.000+00:00
        name: name
        endPoi: endPoi
        id: id
        intermediatePois:
        - intermediatePois
        - intermediatePois
        status: active
      properties:
        id:
          description: GUID generated during Route creation
          type: string
        name:
          type: string
        startPoi:
          description: POI id of the start location for a route
          type: string
        endPoi:
          description: POI id of the end location for a route
          type: string
        intermediatePois:
          items:
            description: POI ids of all the locations which are in the route.
            type: string
          type: array
        status:
          description: Route status to help to exclude or include this in trip planning
          enum:
          - active
          - inactive
          example: active
          type: string
        audit:
          $ref: '#/components/schemas/Audit'
      type: object
      xml:
        name: route
...